+++ tests/test_cli_build.py
@@ -2,12 +2,14 @@ import json
 
 from moncic.runner import UserConfig
 from moncic.unittest import CLITestCase
+from moncic.unittest.sources import Package, SourcesTestCase
 
 
-class CliBuildTests(CLITestCase):
+class CliBuildTests(CLITestCase, SourcesTestCase):
     def test_ci(self) -> None:
+        package = self.get_package("hello")
         self.session.test_simulate_bootstrap("test", {"extends": "rocky8"})
-        res = self.call("monci", "ci", "test")
+        res = self.call("monci", "ci", "test", package.path.as_posix())
         self.assertNoStderr(res)
         with self.match_run_log(self.session.run_log) as m:
             container_log = m.assertPopFirst("test: run container")
@@ -17,7 +19,7 @@ class CliBuildTests(CLITestCase):
                 )
                 cm.assertPopScript("Set up the container filesystem")
                 cm.assertPopScript("Update container packages before build")
-                cm.assertPopScript("Build .")
+                cm.assertPopScript(f"Build {package.path}")
 
         output = json.loads(res.stdout)
         self.assertEqual(
@@ -33,7 +35,7 @@ class CliBuildTests(CLITestCase):
         )
 
         self.assertEqual(output["result"]["artifacts"], [])
-        self.assertEqual(output["result"]["name"], "moncic-ci")
+        self.assertEqual(output["result"]["name"], "hello")
         self.assertTrue(output["result"]["success"])
         self.assertEqual(output["result"]["trace_log"], [])