+++ tests/test_source_distro.py
@@ -9,7 +9,7 @@ from moncic.source.distro import DistroSource, source_types
 from moncic.source.local import Dir, File, Git
 from moncic.source.source import Source
 
-from .source import GitFixture
+from moncic.unittest.source import GitFixture
 
 ROCKY9 = cast(RpmDistro, DistroFamily.lookup_distro("rocky9"))
 SID = cast(DebianDistro, DistroFamily.lookup_distro("sid"))