+++ tests/test_source_remote.py
@@ -6,7 +6,7 @@ from moncic.source import Source
 from moncic.source.local import Git
 from moncic.source.remote import URL
 
-from .source import GitFixture
+from moncic.unittest.source import GitFixture
 
 ROCKY9 = DistroFamily.lookup_distro("rocky9")
 SID = DistroFamily.lookup_distro("sid")